Moved Test environment to the root of the project.
