+++ UploadBOM/__tests__/integration/fixtures/vulnerable-bom.json
