+++ UploadBOM/package.json
@@ -6,9 +6,6 @@
   "scripts": {
     "test": "jest --config=jest.unit.config.js",
     "test:integration": "jest --config=jest.integration.config.js",
-    "docker:start:v4": "DTRACK_MAJOR_VERSION=4 bash __tests__/integration/setup/dtrack-environment.sh start",
-    "docker:start": "DTRACK_MAJOR_VERSION=5 bash __tests__/integration/setup/dtrack-environment.sh start",
-    "docker:stop": "bash __tests__/integration/setup/dtrack-environment.sh stop",
     "build": "babel src --out-dir dist --copy-files"
   },
   "author": "",
